move merge_layers function to graph_modification_utils
